toplev: new scheduler
This is a major change of the group scheduler. It now generates groups
even for objects that do not fit into a single group. It falls back
to grouping by level (thanks to Ahmad for suggesting that) and then
finally splits into smaller groups.

The output parser is now using a global mapping table instead of the
old nested lists.

Also use a better sort mechanism and various other improvements.