toplev: Account duplicated fixed counters
With the SMT model we can have multiple instances of the same fixed
counter in deeper levels, typically with ANY on and ANY off.
The event scheduler didn't correctly account for this, which
can lead to unschedulable groups.

For now promote those to generic counters.

NB still some unschedulable groups left.