simple-pebs: return an error code if PEBS initialization fails
If simple_pebs_cpu_init fails, the module initialization will return 1.
This does not follow convention and causes Linux to emit a warning and
continue loading the module. Unloading the module will then cause it to
try to restore the state saved during loading, which will fail.

Returning an actual error code fixes this.