toplev: Compute MUX from perf data
The MUX metrics in the models never quite worked because it needs
special scheduling. perf instead keeps track of the multiplexing status
of each group and reports it. toplev already tracked in internally
and output it, but not as a separate node which made the data
often disappear when post processing. Reimplement the MUX
node using the perf data.  Also output it by default as part of the perf metrics.
This requires a fairly new perf binary that outputs multiplexing data in CSV
mode